Simple_Caculatror
This C program is a simple calculator that lets users perform basic arithmetic operations: addition, subtraction, multiplication, division, modulus, and exponentiation (power). It displays a menu for selecting an operation and prompts the user to enter two numbers. Based on the chosen operation, it calculates and displays the result. Division and modulus operations include checks to avoid division by zero. The program runs in a loop, allowing multiple calculations until the user chooses to exit.